+++ src/ProjectShopOnline/view/ShopView.java
@@ -1,4 +1,6 @@
-package ProjectShopOnline.model;
+package ProjectShopOnline.view;
+
+import ProjectShopOnline.model.Order;
 
 public interface ShopView {
     void displayOrderDetails(Order order);
